+++ const/ItemIds.lua
@@ -6,8 +6,13 @@ local TOCNAME, BOM = ...
 BOM.ItemId = {}
 
 BOM.ItemId.Mage = {}
+
 BOM.ItemId.Mage.ManaEmerald = 22044
 BOM.ItemId.Mage.ManaRuby = 8008
 BOM.ItemId.Mage.ManaCitrine = 8007
 BOM.ItemId.Mage.ManaJade = 5513
 BOM.ItemId.Mage.ManaAgate = 5514
+
+BOM.ItemId.Warlock = {}
+
+BOM.ItemId.Warlock.SoulShard = 6265