+++ const/SpellIds.lua
@@ -12,6 +12,9 @@ BOM.SpellId.Mage.ConjureManaCitrine = 10053
 BOM.SpellId.Mage.ConjureManaJade = 3552
 BOM.SpellId.Mage.ConjureManaAgate = 759
 
+BOM.SpellId.Warlock = {}
+BOM.SpellId.Warlock.DemonicSacrifice = 18788
+
 BOM.SpellId.Druid = {}
 BOM.SpellId.Druid.TrackHumanoids = 5225